refactor: separate user upload logic
